requirements.txt
streamlit
python-dotenv
transformers
pandas
plotly
google-api-python-client
torch